Vehicle Turning Radius and Misc
Changed variable types to Double for more accurate representation of values
Modified rear wheel speeds to specific ratio to reduce drift and to create a perfect circle.

Misc:
Removed a few osDelays in the code.

+++ Core/Inc/pid.h
@@ -5,13 +5,13 @@
 #include "main.h"
 
 typedef struct{
-	int16_t lastError;
-	int32_t errorIntegral;
+	double lastError;
+	double errorIntegral;
 	int16_t output;
-	int16_t target;
+	double target;
 }pid_instance;
 
-void setTarget(pid_instance *m, int16_t target);
+void setTarget(pid_instance *m, double target);
 //void setSpeed(int16_t speed);
 void setPID(float p, float i, float d);
 void apply_pid(pid_instance *m, int16_t measuredVelocity);
